chore: add api-contract for user-management endpoints

+++ services/auth-service/src/main/java/com/example/auth_service/dto/response/UserManagementResponse.java
@@ -10,6 +10,6 @@ public record UserManagementResponse(
     String role,
     boolean status,
     String profileImageUrl,
-    int eventsOrganized,
-    int eventsAttended
+    long eventsOrganized,
+    long eventsAttended
 ){}
